+++ Sources/Track.swift
@@ -2,7 +2,7 @@ import Foundation
 
 struct Track: Encodable {
     let publishId: String
-    let timestampSecs: Double
+    let timestampSecs: UInt
     let eventType: String
     let deviceId: String
 }
